src/common/requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.2
fnvhash==0.1.0
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
packaging==16.8
peewee==2.10.1
pyparsing==2.2.0
PyYAML==3.12
redis==2.10.5
requests==2.14.2
six==1.10.0
Werkzeug==0.12.2
uwsgi
certbot
